RSpec
The rspec command-line runner understands a `--no-color` flag. This is
not documented per se but it is in the output of `rspec --help`.

It can be passed as a command-line flag, through the `SPEC_OPTS`
environment variable, or in a `.rspec` configuration. Note that a
`~/.rspec` is read first, then `./.rspec`, then `SPEC_OPT`, then
command-line args.