Add unix-shell-script-kit to list of color libraries supporting NO_COLOR
The `unix-shell-script-kit` is a POSIX sh library of utility functions. I'm the maintainer. I just added a NO_COLOR check.

Now, the libary out() function and err() function will skip color when NO_COLOR is set, and also in some other conditions such as if the user is in a dumb terminal.